+++ pkg/distributor/distributor_test.go
@@ -1617,32 +1617,6 @@ func BenchmarkDistributor_Push(b *testing.B) {
 			},
 			expectedErr: "label value too long",
 		},
-		"timestamp too old": {
-			prepareConfig: func(limits *validation.Limits) {
-				limits.RejectOldSamples = true
-				limits.RejectOldSamplesMaxAge = model.Duration(time.Hour)
-			},
-			prepareSeries: func() ([]labels.Labels, []mimirpb.Sample) {
-				metrics := make([]labels.Labels, numSeriesPerRequest)
-				samples := make([]mimirpb.Sample, numSeriesPerRequest)
-
-				for i := 0; i < numSeriesPerRequest; i++ {
-					lbls := labels.NewBuilder(labels.Labels{{Name: model.MetricNameLabel, Value: "foo"}})
-					for i := 0; i < 10; i++ {
-						lbls.Set(fmt.Sprintf("name_%d", i), fmt.Sprintf("value_%d", i))
-					}
-
-					metrics[i] = lbls.Labels()
-					samples[i] = mimirpb.Sample{
-						Value:       float64(i),
-						TimestampMs: time.Now().Add(-2*time.Hour).UnixNano() / int64(time.Millisecond),
-					}
-				}
-
-				return metrics, samples
-			},
-			expectedErr: "timestamp too old",
-		},
 		"timestamp too new": {
 			prepareConfig: func(limits *validation.Limits) {
 				limits.CreationGracePeriod = model.Duration(time.Minute)
@@ -3201,15 +3175,6 @@ func TestDistributorValidation(t *testing.T) {
 				Value:       1,
 			}},
 		},
-		// Test validation fails for very old samples.
-		{
-			labels: []labels.Labels{{{Name: labels.MetricName, Value: "testmetric"}, {Name: "foo", Value: "bar"}}},
-			samples: []mimirpb.Sample{{
-				TimestampMs: int64(past),
-				Value:       2,
-			}},
-			err: httpgrpc.Errorf(http.StatusBadRequest, `timestamp too old: %d metric: "testmetric"`, past),
-		},
 
 		// Test validation fails for samples from the future.
 		{
@@ -3258,8 +3223,6 @@ func TestDistributorValidation(t *testing.T) {
 			flagext.DefaultValues(&limits)
 
 			limits.CreationGracePeriod = model.Duration(2 * time.Hour)
-			limits.RejectOldSamples = true
-			limits.RejectOldSamplesMaxAge = model.Duration(24 * time.Hour)
 			limits.MaxLabelNamesPerSeries = 2
 
 			ds, _, _ := prepare(t, prepConfig{